Background

Description

CosmoTools provides a set of functions for the background cosmology.

Structures

CosmoTools.Baryons — Type

Baryons <: Matter

CosmoTools.BkgCosmology — Type

BkgCosmology{T<:Real}

Abstract type: generic background cosmology

CosmoTools.ColdDarkMatter — Type

ColdDarkMatter <: Matter

CosmoTools.Curvature — Type

Curvature <: Species

CosmoTools.DarkEnergy — Type

DarkEnergy <: Species

CosmoTools.FLRW — Type

FLRW(h, Ω_χ0, Ω_b0; T0_CMB_K = 2.72548, Neff = 3.04)

Creates a FLRW instance

Arguments

h::Real : hubble parameter (dimensionless)

Ω_χ0::Real: cold dark matter abundance today (dimensionless)

Ω_b0::Real: baryon abundance today (dimensionless)

T0_CMB_K::Real: temperature of the CMB today (in Kelvin)

Neff::Real: effective number of neutrinos (dimensionless)

Returns

A FLRW structure

CosmoTools.FLRW — Type

FLRW{T<:Real} <: BkgCosmology{T}

Defines a flat FLRW background cosmology

CosmoTools.Matter — Type

Matter <: Species

CosmoTools.Neutrinos — Type

Neutrinos <: Radiation

CosmoTools.Photons — Type

Photons <: Radiation

CosmoTools.Radiation — Type

Radiation <: Species

CosmoTools.Species — Type

Species

Abstract type: generic species populating the Universe

Constants

The module provides some predefined background cosmologies

CosmoTools.edsPlanck18_bkg — Constant

edsplanck18_bkg::FLRW  = FLRW(0.6736, 0.3, 0)

Flat FLRW cosmology with Planck18 parameters and no baryons

CosmoTools.planck18_bkg — Constant

planck18_bkg::FLRW = FLRW(0.6736, 0.26447, 0.04930)

Flat FLRW cosmology with Planck18 parameters

Functions

CosmoTools.a_to_z — Method

a_to_z(z)

Convert scale parameter to redshift. Simple implementation of

$z = \frac{1}{a}-1$

Arguments

a::Real : scale factor (dimensionless)

CosmoTools.convert_cosmo — Method

Convert bkg_cosmo object attributes to another type

CosmoTools.growth_factor — Function

growth_factor(z, bkg_cosmo)

Exact growth factor in a matter-Λ Universe computed from an integral
Carroll et al. 1992 (Mo and White p. 172)
Corresponds to D1(a=1) with the definition of Dodelson 2003

Arguments

z::Real : redshift (dimensionless)

bkg_cosmo::BkgCosmology : background cosmology, default planck18_bkg

Kwargs

Any argument that can be passed to `QuadGK.quadgk'

CosmoTools.growth_factor_Carroll — Function

growth_factor_Carroll(z, bkg_cosmo)

Approximate growth factor in a matter-Λ Universe (faster than growth_factor)
Carroll et al. 1992 (Mo and White p. 172)
Corresponds to D1(a=1) with the definition of Dodelson 2003

Arguments

z::Real : redshift (dimensionless)

bkg_cosmo::BkgCosmology : background cosmology, default planck18_bkg

CosmoTools.hubble_E — Function

hubble_E(z, bkg_cosmo)

Hubble evolution function E(z) (dimensionless) at redshift z.

Arguments

z::Real: Redshift (default: 0.0)

bkg_cosmo::BkgCosmology: Cosmological background model (default: planck18_bkg)

CosmoTools.hubble_E2 — Function

hubble_E2(z, bkg_cosmo)

Squared Hubble evolution function E²(z) (dimensionless) at redshift z.

Arguments

z::Real: Redshift (default: 0.0)

bkg_cosmo::BkgCosmology: Cosmological background model (default: planck18_bkg)

CosmoTools.hubble_H — Function

hubble_H(z, bkg_cosmo)

Hubble rate H(z) (in km/s/Mpc) of the Universe at redshift z.

Arguments

z::Real: Redshift (default: 0.0)

bkg_cosmo::BkgCosmology: Cosmological background model (default: planck18_bkg)

CosmoTools.hubble_H0 — Function

hubble_H0(bkg_cosmo)

Hubble constant H₀ (in km/s/Mpc) for the cosmology.

Arguments

bkg_cosmo::BkgCosmology: Cosmological background model (default: planck18_bkg)

CosmoTools.hubble_H0_s — Function

hubble_H0_s(bkg_cosmo)

Hubble constant H₀ (in 1/s) for the cosmology.

Arguments

bkg_cosmo::BkgCosmology: Cosmological background model (default: planck18_bkg)

CosmoTools.k_eq_mr — Method

k_eq_mr(bkg_cosmo)

Wave number at matter–radiation equality.

Arguments

bkg_cosmo::BkgCosmology: Cosmological background model

CosmoTools.lookback_redshift — Function

lookback_lookback_redshift(t, bkg_cosmo; kws...)

lookback redshift of the Universe for t in (s)

Arguments

t::Real : time (s)

bkg_cosmo::BkgCosmology : background cosmology (dimensionless), default planck18_bkg

Kwargs

Any argument that can be passed to `QuadGK.quadgk'

CosmoTools.lookback_time — Function

lookback_time(z, bkg_cosmo; kws...)

lookback time of the Universe (s)

Arguments

z::Real : redshift (dimensionless)

bkg_cosmo::BkgCosmology : background cosmology (dimensionless), default planck18_bkg

Kwargs

Any argument that can be passed to `QuadGK.quadgk'

CosmoTools.mean_ρ — Method

mean_ρ(T, z = 0, bkg_cosmo = planck18_bkg) where {T<:Species}

Energy density of species T at redshift z and for the background cosmology bkg_cosmo

CosmoTools.temperature_CMB_K — Function

temperature_CMB_K(z, bkg_cosmo)

CMB temperature (in K) of the Universe at redshift z.

Arguments

z::Real: Redshift (default: 0.0)

bkg_cosmo::BkgCosmology: Cosmological background model (default: planck18_bkg)

CosmoTools.universe_age — Function

universe_age(z, bkg_cosmo; kws...)

age of the universe (s)

Arguments

z::Float64 : redshift (dimensionless), default 0.0

bkg_cosmo::BkgCosmology : background cosmology (dimensionless), default planck18_bkg

Kwargs

Any argument that can be passed to `QuadGK.quadgk'

CosmoTools.z_to_a — Method

z_to_a(z)

Convert redshift z to scale parameter a. Simple implementation of

$a = \frac{1}{1+z}$

Arguments

z::Real : redshift (dimensionless)

CosmoTools.Ω — Method

Ω(T, z = 0, bkg_cosmo = planck18_bkg) where {T<:Species}

Abundance of species T at redshift z and for the background cosmology bkg_cosmo

CosmoTools.ρ_critical — Function

Critical density (in Msun/Mpc^3) of the Universe at redshift z (by default z=0) for the cosmology bkg_cosmo