Transfer Function

Description

In CosmoTools you can define a transfer function to be applied to the primordial curvature power spectrum in order to obtain the matter power spectrum in the matter dominater era. Some usual transfer functions are already included.

Structures

Constants

Functions

CosmoTools.transfer_function — Method

transfer_function(k, [bkg_cosmo::BkgCosmology], with_baryons = true)

If bkg_cosmo is given computes the transfer function for